requirements.txt
Flask==2.1.1
Authlib==1.0.0
python-dotenv==1.0.0
